step5 refactoring
