+++ src/main/java/racingCar/view/InputView.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.view;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -13,13 +13,13 @@ public class InputView {
         return scanner.nextInt();
     }
 
-    public String[] nameList() {
+    public static String[] nameList() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         return scanner.nextLine().split(",");
 
     }
 
-    public int numberOfGame() {
+    public static int numberOfGame() {
         System.out.println("시도할 회수는 몇 회 인가요?");
         return scanner.nextInt();
     }
