+++ src/test/java/racingCar/NumberOfGameTest.java
@@ -1,7 +1,7 @@
 package racingCar;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingCar.domain.NumberOfGame;
 
 import static org.assertj.core.api.Assertions.*;
