+++ src/test/java/racingCar/RandomNumberTest.java
@@ -2,6 +2,7 @@ package racingCar;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingCar.domain.RandomNumber;
 
 public class RandomNumberTest {
     @Test
